global object ident e2e
