+++ e2e/global-object-identification/gateway.config.ts
@@ -1,0 +1,5 @@
+import { defineConfig } from '@graphql-hive/gateway';
+
+export const gatewayConfig = defineConfig({
+  globalObjectIdentification: true,
+});
