--- e2e/relay-additional-resolvers/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "@e2e/relay-additional-resolvers",
-  "private": true,
-  "dependencies": {
-    "@apollo/subgraph": "^2.10.0",
-    "graphql": "^16.9.0",
-    "graphql-sse": "^2.5.3"
-  }
-}
